+++ presentation/src/main/java/com/stop/ui/mission/MissionFragment.kt
@@ -2,7 +2,10 @@ package com.stop.ui.mission
 
 import android.Manifest
 import android.animation.Animator
+import android.content.BroadcastReceiver
+import android.content.Context
 import android.content.Intent
+import android.content.IntentFilter
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
@@ -10,6 +13,7 @@ import android.view.ViewGroup
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.snackbar.Snackbar
@@ -23,18 +27,21 @@ import com.stop.model.Location
 import com.stop.model.MissionStatus
 import com.stop.ui.alarmsetting.AlarmSettingViewModel
 import com.stop.ui.mission.MissionService.Companion.MISSION_LAST_TIME
+import com.stop.ui.mission.MissionService.Companion.MISSION_LOCATIONS
 import com.stop.ui.mission.MissionService.Companion.MISSION_OVER
 import com.stop.ui.util.Marker
+import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectIndexed
 import kotlinx.coroutines.launch
 
+@AndroidEntryPoint
 class MissionFragment : Fragment(), MissionHandler {
 
     private var _binding: FragmentMissionBinding? = null
     private val binding: FragmentMissionBinding
         get() = _binding!!
 
-    private val missionViewModel: MissionViewModel by activityViewModels()
+    private val missionViewModel: MissionViewModel by viewModels()
     private val alarmSettingViewModel: AlarmSettingViewModel by activityViewModels()
 
     private lateinit var tMap: MissionTMap
@@ -47,6 +54,7 @@ class MissionFragment : Fragment(), MissionHandler {
         super.onCreate(savedInstanceState)
 
         setMissionService()
+        setBroadcastReceiver()
         missionViewModel.missionStatus.value = MissionStatus.ONGOING
     }
 
@@ -88,6 +96,24 @@ class MissionFragment : Fragment(), MissionHandler {
         } else {
             requireActivity().startService(missionServiceIntent)
         }
+    }
+
+    private fun setBroadcastReceiver() {
+        val intentFilter = IntentFilter().apply {
+            addAction(MissionService.MISSION_USER_INFO)
+        }
+
+        val receiver = object : BroadcastReceiver() {
+            override fun onReceive(context: Context?, intent: Intent?) {
+                intent?.let {
+                    missionViewModel.lastTime.value = intent.getStringExtra(MISSION_LAST_TIME)
+                    missionViewModel.userLocations.value =
+                        intent.getParcelableArrayListExtra<Location>(MISSION_LOCATIONS) as ArrayList<Location>
+                }
+            }
+        }
+
+        requireActivity().registerReceiver(receiver, intentFilter)
     }
 
     private fun setTimer() {
